refactor: changed to one function, one export.
