+++ library/src/utils/checkDateInRange.ts
@@ -1,0 +1,2 @@
+export const checkDateInRange = (startDate: Date, date: Date, endDate: Date) =>
+  startDate <= date && date <= endDate;
