--- library/src/utils/css.ts
@@ -1,7 +1,0 @@
-const getClassNameByType = (
-  styles: { [key: string]: string },
-  className: string,
-  type?: string,
-) => `${styles[className]} ${type ? styles[type.toLowerCase()] : ''}`;
-
-export { getClassNameByType };
