+++ library/src/utils/formatHourAndMinutes.ts
@@ -1,0 +1,8 @@
+export const formatHourAndMinutes = (data: Date) => {
+  const hours = data.getHours();
+  const minutes = data.getMinutes();
+  const minutesFormat = minutes < 10 ? `0${minutes}` : minutes;
+  const currentTime = minutes === 0 ? hours : `${hours}:${minutesFormat}`;
+
+  return currentTime; // HH:MM
+};
