+++ library/src/utils/getTodayFromTime.ts
@@ -1,0 +1,19 @@
+export const getTodayFromTime = (
+  hours: number,
+  minutes: number,
+  second: number,
+) => {
+  const now = new Date();
+  const year = now.getFullYear();
+  const month = (now.getMonth() + 1).toString().padStart(2, '0'); // Months are zero-indexed
+  const day = now.getDate().toString().padStart(2, '0');
+
+  const yearMonthDay = `${year}-${month}-${day}`;
+  const hourFormat = hours < 10 ? `0${hours}` : hours;
+  const minutesFormat = minutes < 10 ? `0${minutes}` : minutes;
+  const secondeFormat = second < 10 ? `0${second}` : second;
+
+  return new Date(
+    `${yearMonthDay}T${hourFormat}:${minutesFormat}:${secondeFormat}`,
+  );
+};
