+++ library/src/utils/hoursToMilliseconds.ts
@@ -1,0 +1,2 @@
+export const hoursToMilliseconds = (hours: number): number =>
+  hours * 60 * 60 * 1000;
