+++ library/src/utils/minutesToMilliseconds.ts
@@ -1,0 +1,2 @@
+export const minutesToMilliseconds = (minutes: number): number =>
+  minutes * 60 * 1000;
