+++ library/src/utils/position/getTaskVerticalSizeStyle.ts
@@ -1,5 +1,6 @@
 import { BaseTask } from '../../types/baseTask';
-import { calculateTargetPosition } from '../date/calculation';
+import { calculateTargetHeightPercent } from '../calculateTargetHeightPercent';
+import { calculateTargetPosition } from '../calculateTargetPosition';
 
 export default function getTaskVerticalSizeStyle<T extends BaseTask>(
   task: T,
@@ -9,13 +10,20 @@ export default function getTaskVerticalSizeStyle<T extends BaseTask>(
   const { startPercent, endPercent } = calculateTargetPosition(
     startTime,
     endTime,
-    task.startTime!,
-    task.endTime!,
+    task.startTime,
+    task.endTime,
   );
+
+  const taskHeightPercent = calculateTargetHeightPercent({
+    startTime,
+    endTime,
+    targetStartTime: task.startTime,
+    targetEndTime: task.endTime,
+  });
 
   const taskStyle: React.CSSProperties = {
     top: `${startPercent}%`,
-    height: `${endPercent}%`,
+    height: `${taskHeightPercent}%`,
   };
 
   return taskStyle;
