+++ library/src/utils/secondsToMilliseconds.ts
@@ -1,0 +1,2 @@
+export const secondsToMilliseconds = (seconds: number): number =>
+  seconds * 1000;
